Refactor home to use sub components
