+++ frontend/test/unit/specs/Util/CopyableText.spec.js
