+++ frontend/test/unit/specs/Util/Spinner.spec.js
